+++ src/controllers/admin/BulkShipmentLabelsController.php
@@ -1,27 +1,4 @@
 <?php
-/**
- * 2019 Packlink
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Academic Free License (AFL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/afl-3.0.php
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@prestashop.com so we can send you a copy immediately.
- *
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade PrestaShop to newer
- * versions in the future. If you wish to customize PrestaShop for your
- * needs please refer to http://www.prestashop.com for more information.
- *
- * @author    Packlink <support@packlink.com>
- * @copyright 2019 Packlink Shipping S.L
- * @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
- */
 
 use Logeecom\Infrastructure\Logger\Logger;
 use Logeecom\Infrastructure\ORM\RepositoryRegistry;
@@ -84,6 +61,7 @@ class BulkShipmentLabelsController extends ModuleAdminController
         $orderIds = \Tools::getValue('orders');
 
         $tmpDirectory = _PS_MODULE_DIR_ . 'packlink/tmp';
+        /** @noinspection MkdirRaceConditionInspection */
         if (!empty($orderIds) && !is_dir($tmpDirectory) && !mkdir($tmpDirectory)) {
             throw new \RuntimeException(
                 TranslationUtility::__('Directory "%s" was not created', array($tmpDirectory))
@@ -104,6 +82,7 @@ class BulkShipmentLabelsController extends ModuleAdminController
             }
 
             $bulkLabelDirectory = _PS_MODULE_DIR_ . 'packlink/labels';
+            /** @noinspection MkdirRaceConditionInspection */
             if (!is_dir($bulkLabelDirectory) && !mkdir($bulkLabelDirectory)) {
                 throw new \RuntimeException(
                     TranslationUtility::__('Directory "%s" was not created', array($bulkLabelDirectory))