+++ src/override/controllers/admin/AdminOrdersController.php
@@ -1,27 +1,4 @@
 <?php
-/**
- * 2019 Packlink
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Academic Free License (AFL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/afl-3.0.php
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@prestashop.com so we can send you a copy immediately.
- *
- * DISCLAIMER
- *
- * Do not edit or add to this file if you wish to upgrade PrestaShop to newer
- * versions in the future. If you wish to customize PrestaShop for your
- * needs please refer to http://www.prestashop.com for more information.
- *
- * @author    Packlink <support@packlink.com>
- * @copyright 2019 Packlink Shipping S.L
- * @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
- */
 
 /** @noinspection PhpUnusedPrivateMethodInspection */
 /** @noinspection PhpIncludeInspection */
@@ -34,7 +11,7 @@ class AdminOrdersController extends AdminOrdersControllerCore
     /** OLD PART START */
     /**
      * ***** everything inside this block will be removed on install/upgrade ****
-     * this is a cleanup of the overrides added in version 2.0.0.
+     * This is a cleanup of the overrides added previous version(s).
      */
     const PACKLINK_ORDER_DRAFT_TEMPLATE = '';
     const PACKLINK_ORDER_ICONS_TEMPLATE = '';
@@ -82,7 +59,7 @@ class AdminOrdersController extends AdminOrdersControllerCore
      */
     public function printPDFIcons($orderId, $tr)
     {
-        return $this->packlinkAdminOrderController->printPDFIcons($orderId, $this->context);
+        return $this->packlinkAdminOrderController->renderPdfIcons($orderId, $this->context);
     }
 
     /**